CRUD methods for Reviews

+++ travelAgency/src/main/java/com/example/travelAgency/repository/ReviewsRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface ReviewsRepository extends JpaRepository<Reviews,Integer> {
+public interface ReviewsRepository extends JpaRepository<Reviews,Long> {
 }
